Move images.py to image.py
This is more parallel using module names as adjectives, eg math properties,
nested properties, image properties
